Add OTel log emission, split main.go, update to current LLM models
- Emit correlated OTel log records (ERROR/WARN/INFO) via OTLP alongside
  traces, with per-service LoggerProvider and span context correlation
- Add -no-logs flag to disable log emission
- Split monolithic main.go (3600+ lines) into 4 files: main.go (setup),
  scenarios.go (traditional flows), scenarios_ai.go (AI flows), helpers.go
- Replace hardcoded gpt-4o/gpt-4o-mini with randomized model registry
  across providers (OpenAI gpt-5.x, Anthropic claude-*-4-5, DeepSeek,
  Grok, Mistral) with correct gen_ai.system and endpoint per provider
